Document Terraform examples (#167)
* Optional Resource Group creation in adb-lakehouse module

The commit contains the implementation for the workspace resource group. However this change requires to no more use the local.rg_location variable, since the value is known after the apply, and rhis force the replacement of all of the resources

* Fix resource group name value in the azure data factory resource

* Document examples

Most of README files were already defined. TFDocs updated in each of them

* Remove modules reference and change how to use guide