Fix incorrect use of `location` in `adb-lakehouse` module (#182)
When `create_resource_group` is set to `false` we trued to use the location of existing
resource group, but it wasn't complete as some resources still used `location` passed as
variable.

With these changes we can omit `location` when we don't create a resource group, so we'll
use default location from resource group itself.  But if we don't create a resource group
and `location` is set to non-empty value, then all resources will be created in that
location.

Also fixed specification of managed resource group for Databricks workspace.

Resolves #181