Databricks on GCP data exfiltration protection workspace deployment (#172)
* Optional Resource Group creation in adb-lakehouse module

The commit contains the implementation for the workspace resource group. However this change requires to no more use the local.rg_location variable, since the value is known after the apply, and rhis force the replacement of all of the resources

* Fix resource group name value in the azure data factory resource

* Document examples

Most of README files were already defined. TFDocs updated in each of them

* Remove modules reference and change how to use guide

* Databricks on GCP data exfiltration protection workspace deployment

* Align implementation with GCE (CMv1) version

Starting from provider version 1.71 CMv1 is supported for Databricks on GCP

* Remove references to GKE related variables

* Add utility for Unity Catalog setup in GCP

This is an initial implementation. I will enhance it in the future commits in order to include metastore admin assignment, workspace-metastore binding, catalog owner, catalog-workspace binding.
In case the module can be built in order to be cloud agnostic.

* Add missing firewall rule + Fix architecture diagram and README + Run terraform fmt